Rename neovim to nvim
